more refactors and cleanups
